pytorch/requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
PuLP>=2.6.0
scikit-image>=0.19.0
scipy>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.64.0
streamlit>=1.25.0
Pillow>=9.0.0
matplotlib>=3.5.0
